casting casting casting people!
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@22677 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/commands.cxx
@@ -1843,7 +1843,7 @@ bool HandicapCommand::operator() (const char	 *,
   int t = playerData->getIndex();
 
   if (clOptions->gameOptions & HandicapGameStyle) {
-      const int maxhandicap =  BZDB.eval(StateDatabase::BZDB_HANDICAPSCOREDIFF);
+      const int maxhandicap =  (int)BZDB.eval(StateDatabase::BZDB_HANDICAPSCOREDIFF);
       for (int i = 0; i < curMaxPlayers; i++) {
 	GameKeeper::Player *p = GameKeeper::Player::getPlayerByIndex(i);
 	if (p != NULL) {
